feat: adds spinner element
The spinner is an object that lets the user know the app is fetching data.

There is a HTML div that can be stylized as any kind of loading indicator.
(For this initial commit, just "Loading..." is displayed.)

In addition to the HTML, the spinner keeps private state so many
asynchronous events can request the spinner and it will wait until all
the requesters have released the spinner before disappearing.